backend/requirements.txt
appdirs==1.4.4
asttokens==2.4.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.7.0
Brotli==1.0.9
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.0
executing==2.0.1
flake8==4.0.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar @ git+https://github.com/pallets-eco/flask-debugtoolbar@9b63ad1837458f14597b87ad266da3d38835071f
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.29.1
forex-python==1.8
fs==2.4.12
gast==0.5.2
greenlet==3.0.3
html5lib==1.1
httpie==2.6.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
more-itertools==8.10.0
mpmath==0.0.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==23.2
parso==0.8.3
pep8==1.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-dotenv==1.0.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-toolbelt==0.9.1
scipy==1.8.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.9
systemd-python==234
toml==0.10.2
traitlets==5.14.0
typing_extensions==4.9.0
ufoLib2==0.13.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.1.0
wadllib==1.3.6
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==2.3.8
WTForms==3.1.2
zipp==1.0.0
